requirement.txt
# Core graph + orchestration
langgraph==0.0.55

# LangChain core ecosystem (0.2.x – stable)
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16

# Groq LLM integration
langchain-groq==0.1.9

# Web search (SerpAPI)
google-search-results==2.4.2

# Environment variables
python-dotenv==1.0.1

# Web scraping
requests==2.32.3
beautifulsoup4==4.12.3

# Observability / tracing
langfuse==2.44.0
streamlit \